refactor: GameManager를 console의 상위로 이동
