+++ src/main/java/chess/console/command/Start.java
@@ -1,6 +1,6 @@
 package chess.console.command;
 
-import chess.console.GameManager;
+import chess.GameManager;
 import chess.console.view.OutputView;
 
 public final class Start implements Command{
